Moved frame pos generator to Area
