+++ src/main/scala/com/yogpc/qp/machines/quarry/Target.java
@@ -6,8 +6,6 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Objects;
 import java.util.Set;
-import java.util.function.Function;
-import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
 import com.yogpc.qp.machines.Area;
@@ -180,24 +178,15 @@ final class FrameTarget extends Target {
 
     FrameTarget(Area area) {
         this.area = area;
-        this.iterator = getPosStream(area).iterator();
+        this.iterator = Area.getFramePosStream(area).iterator();
         this.currentTarget = iterator.hasNext() ? iterator.next() : null;
     }
 
     FrameTarget(Area area, BlockPos pre) {
         this.area = area;
-        this.iterator = getPosStream(area)
+        this.iterator = Area.getFramePosStream(area)
             .dropWhile(p -> !p.equals(pre)).iterator();
         this.currentTarget = iterator.hasNext() ? iterator.next() : null;
-    }
-
-    @Nonnull
-    private Stream<BlockPos> getPosStream(Area area) {
-        return Stream.of(
-            makeSquare(area, area.minY()),
-            makePole(area, area.minY() + 1, area.maxY()),
-            makeSquare(area, area.maxY())
-        ).flatMap(Function.identity());
     }
 
     @Override
@@ -217,7 +206,7 @@ final class FrameTarget extends Target {
     @Override
     @Nonnull
     public Stream<BlockPos> allPoses() {
-        return getPosStream(area);
+        return Area.getFramePosStream(area);
     }
 
     @Override
@@ -233,24 +222,6 @@ final class FrameTarget extends Target {
 
     static FrameTarget from(CompoundTag tag) {
         return new FrameTarget(Area.fromNBT(tag.getCompound("area")).orElseThrow(), BlockPos.of(tag.getLong("currentTarget")));
-    }
-
-    static Stream<BlockPos> makeSquare(Area area, int y) {
-        return Stream.of(
-            IntStream.rangeClosed(area.minX(), area.maxX()).mapToObj(x -> new BlockPos(x, y, area.minZ())), // minX -> maxX, minZ
-            IntStream.rangeClosed(area.minZ(), area.maxZ()).mapToObj(z -> new BlockPos(area.maxX(), y, z)), // maxX, minZ -> maxZ
-            IntStream.iterate(area.maxX(), x -> x >= area.minX(), x -> x - 1).mapToObj(x -> new BlockPos(x, y, area.maxZ())), // maxX -> minX, maxZ
-            IntStream.iterate(area.maxZ(), z -> z >= area.minZ(), z -> z - 1).mapToObj(z -> new BlockPos(area.minX(), y, z)) // minX, maxZ -> minZ
-        ).flatMap(Function.identity());
-    }
-
-    static Stream<BlockPos> makePole(Area area, int yMin, int yMax) {
-        return IntStream.rangeClosed(yMin, yMax).boxed().flatMap(y ->
-            Stream.of(
-                new BlockPos(area.minX(), y, area.minZ()),
-                new BlockPos(area.maxX(), y, area.minZ()),
-                new BlockPos(area.maxX(), y, area.maxZ()),
-                new BlockPos(area.minX(), y, area.maxZ())));
     }
 
     @Override
